refactor: 연속되는 step 을 체크하는 로직중복 제거
