+++ src/main/java/ladder/domain/LadderLevel.java
@@ -2,6 +2,7 @@ package ladder.domain;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.function.BiConsumer;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
@@ -20,25 +21,13 @@ public class LadderLevel {
     }
 
     private void shouldNotConsecutivelyStep(List<LevelItem> items) {
-        IntStream.range(0, items.size())
-                .filter(idx -> items.get(idx).isStep())
-                .forEach(idx -> {
-                    if (nextStep(items, idx).isPresent())
-                        throw new IllegalArgumentException("step 은 연속해서 구성할 수 없습니다");
-                });
+        new ConsecutivelyStepChecker(items).check((levelItems, idx) -> {
+            throw new IllegalArgumentException("step 은 연속해서 구성할 수 없습니다");
+        });
     }
 
     private void shouldBar(LevelItem first, LevelItem last) {
         if (!first.isBar() || !last.isBar()) throw new IllegalArgumentException("처음과 끝은 Bar 로 구성해야 합니다.");
-    }
-
-    private Optional<LevelItem> nextStep(List<LevelItem> items, int idx) {
-        int nextStepIdx = idx + 2;
-        if (items.size() > nextStepIdx
-                && items.get(nextStepIdx).isStep()) {
-            return Optional.ofNullable(items.get(nextStepIdx));
-        }
-        return Optional.ofNullable(null);
     }
 
     @Override
