+++ src/main/java/ladder/domain/LadderLevelGenerator.java
@@ -1,7 +1,6 @@
 package ladder.domain;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.Random;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -26,22 +25,11 @@ public interface LadderLevelGenerator {
                         .collect(Collectors.toList());
 
                 // 연속되는 step 제거
-                IntStream.range(0, size)
-                        .filter(idx -> items.get(idx).isStep())
-                        .forEach(idx -> {
-                            nextStep(items, idx).ifPresent(nextStep -> items.set(idx, LevelItem.emptyStep()));
-                        });
+                new ConsecutivelyStepChecker(items).check((levelItems, idx) -> {
+                    levelItems.set(idx, LevelItem.emptyStep());
+                });
 
                 return items;
-            }
-
-            private Optional<LevelItem> nextStep(List<LevelItem> items, int idx) {
-                int nextStepIdx = idx + 2;
-                if (items.size() > nextStepIdx
-                        && items.get(nextStepIdx).isStep()) {
-                    return Optional.ofNullable(items.get(nextStepIdx));
-                }
-                return Optional.ofNullable(null);
             }
         };
     }
