+++ src/test/java/ladder/domain/LadderLevelGeneratorTest.java
@@ -1,7 +1,5 @@
 package ladder.domain;
 
-import ladder.domain.LadderLevelGenerator;
-import ladder.domain.LevelItem;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
